Add curie query support (mychem version)  (#34)
* Add the biolink prefix mapping to the settings ...

Also separate the settings into a separate file for isolating the
configuration from client construction methods

* Add work for supporting CURIE query support ...

Specifically targets _getannotation and _getannotations methods at the
moment and only for the mygene instance

* Update the tox configuration to use more modern python

* Change the default value for the prefix mapping

* Add proper list formatting for the fields support

* Add proper multiple id/query argument handling

* Add expectedFailure decorators to the chem tests

* Remove the default the regex pattern matching ...

Not all of the different sites will work with that. Ran into issues
with that default pattern with myvariant. Instead we should view this
feature as specifically targetting support for the biolink prefixes
and nothing else

* Remove client side CURIE support ...

Pivot to adding testing and verification that the API / server side
implementation supports the CURIE ID

* Add the CURIE ID query testing for the mygene instance

* Remove unused pytest import

* Add the curie id tests with mychem as the target

* Revert modifications to the settings configuration

* Update the pyproject file

---------

Co-authored-by: Johnathan Schaff <jschaff@scripps.edu>