Add custom proxy mount creation for the cache clients ...
Due to the httpx client constructor design, if we pass
and HTTPTransport, then it assumes the proxy has been configured
and will not attempt to discover it from the environment.
Therefore we have to do that manually for the cache clients
as hishel requires an HTTPTransport. We due this through the
mounts property provided by the client constructor