Bugfix client proxy autodiscovery (#40)
* Remove the HTTPTransport from the client building ...

Adding a blank transport removes our auto-discovery of the
proxy configurations. Removing it for the base level tests
appears to allow for automatic proxy configuration discovery

* Re-add the alwayslist utility function

* Add custom proxy mount creation for the cache clients ...

Due to the httpx client constructor design, if we pass
and HTTPTransport, then it assumes the proxy has been configured
and will not attempt to discover it from the environment.
Therefore we have to do that manually for the cache clients
as hishel requires an HTTPTransport. We due this through the
mounts property provided by the client constructor

* Disable client request timeouts ...

I don't believe this will be permanent behavior but due
to the transition from 0.3.1 -> 0.4.0 with the new HTTP library,
I want to maintain as much backwards compatibility for the moment
so that behavior doesn't suddenly change for users. In the future
I'd likely set the timeout to around 60 seconds with the option
for user configuration upon client creation to allow customization

---------

Co-authored-by: jschaff <jschaff@scripps.edu>